cdc-stream-processor/cdc_stream_processor.py
import json

try:
    from kafka import KafkaProducer, KafkaConsumer
except ImportError:
    # Mock implementations for development/testing when kafka is not available
    class KafkaProducer:
        def __init__(self, bootstrap_servers):
            self.bootstrap_servers = bootstrap_servers
            print(f"Mock KafkaProducer initialized with {bootstrap_servers}")
            
        def send(self, topic, value):
            print(f"Mock sending to {topic}: {value}")
            
    class KafkaConsumer:
        def __init__(self, topic, bootstrap_servers):
            self.topic = topic
            self.bootstrap_servers = bootstrap_servers
            print(f"Mock KafkaConsumer initialized for {topic} with {bootstrap_servers}")
            
        def __iter__(self):
            return self
            
        def __next__(self):
            # This will make the consumer stop after one iteration in mock mode
            raise StopIteration

class CDCStream:
    def __init__(self, bootstrap_servers):
        self.producer = KafkaProducer(bootstrap_servers=bootstrap_servers)
        self.consumer = KafkaConsumer('cdc_topic', bootstrap_servers=bootstrap_servers)

    def stream_changes(self, changes):
        for change in changes:
            self.producer.send('cdc_topic', value=str(change).encode())

    def process_changes(self):
        for message in self.consumer:
            change = json.loads(message.value.decode())
            print(f"Processing change: {change}")
            # Logic to apply changes to the Raw Data Vault
