cdc-stream-processor/requirements.txt
# Requirements for CDC Stream Processor
kafka-python>=2.0.2
requests>=2.28.1
